Creat a authMiddleware to protect 'workouts' routes & applied in App.jsx, WorkoutForm.jsx & WorkoutDetails.jsx

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,17 +1,25 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext.js'
 import { format } from "@formkit/tempo"
 import { Toaster, toast } from "sonner"
+import useAuthContext from "../hooks/useAuthContext"
 
 const WorkoutDetails = ({ workout }) => {
   
   const { dispatch } = useWorkoutsContext()
-
   const dateReadable = format(workout.createdAt, "full")
+  const { user } = useAuthContext()
 
   const handleDelete = async () => {
 
+    if(!user) {
+      return
+    }
+
     const response = await fetch('http://localhost:3000/api/workouts/'+workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
 
     const json = await response.json()
